requirements_production.txt
# Dependências essenciais para produção
Django==5.2.3
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
django-ckeditor-5==0.2.11
django-autocomplete-light==3.9.1
Pillow==11.3.0
qrcode==8.2
reportlab==4.4.2 
PyPDF2>=3.0.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.0
pytz==2024.1
python-dateutil==2.9.0
